+++ wp-includes/class-wp-walker.php
@@ -310,7 +310,7 @@ class Walker {
 			$start  = ( (int) $page_num - 1 ) * (int) $per_page;
 			$end    = $start + $per_page;
 			if ( -1 == $max_depth ) {
-				$this->max_pages = ceil( $total_top / $per_page );
+				$this->max_pages = (int) ceil( $total_top / $per_page );
 			}
 		}
 
@@ -354,7 +354,7 @@ class Walker {
 
 		$total_top = count( $top_level_elements );
 		if ( $paging ) {
-			$this->max_pages = ceil( $total_top / $per_page );
+			$this->max_pages = (int) ceil( $total_top / $per_page );
 		} else {
 			$end = $total_top;
 		}
