+++ wp-includes/rest-api/endpoints/class-wp-rest-comments-controller.php
@@ -300,7 +300,7 @@ class WP_REST_Comments_Controller extends WP_REST_Controller {
 			$prepared_args['orderby'] = 'none';
 
 			$total_comments = $query->query( $prepared_args );
-			$max_pages      = ceil( $total_comments / $request['per_page'] );
+			$max_pages      = (int) ceil( $total_comments / $request['per_page'] );
 		}
 
 		$response = rest_ensure_response( $comments );
